Adding all code to trait.
